Basic workload config seems to init correctly and guest-init is idempotent. Rebuild should take ~2min.

+++ fedora_bootstrap.sh
@@ -1,17 +1,26 @@
 #!/bin/bash
 
-# This will install everything needed to run benchmarks on Fedora
-dnf install -y \
-  libcgroup-tools \
-  time \
-  mpich-devel \
-  openblas-devel \
-  python-numpy
+# Just checking if stuff is installed can take forever on fedora and
+# fedora_bootstrap might be called multiple times. We want it fast the second
+# time.
+if [ ! -e /PFA_INIT ]; then
+  echo "Installing required packages"
+  # This will install everything needed to run benchmarks on Fedora
+  dnf install -y \
+    libcgroup-tools \
+    time \
+    mpich-devel \
+    openblas-devel \
+    python-numpy
 
-pip install \
-  spambayes
+  pip install \
+    spambayes
 
-cd /root
-./build.sh
+  echo "PFA packages installed" > /PFA_INIT
+fi
+
+echo "building benchmarks"
+cd /root/benchmarks
+build.sh
 
 poweroff
